requirements.txt
astroquery
astropy
Pillow
numpy==1.26.4
matplotlib
scipy
scikit-image